refactor: split download.py into poll / processing / transfer modules (#146)
Post-resolver-deletion split, sequenced so only surviving code moves.
Pure moves — no behaviour change:

- lib/download.py (~980 lines) keeps the poll state machine and
  orchestration: DownloadDB protocol, poll_active_downloads,
  _poll_one_active_download, the completed-processing/timeout/retry
  transitions, ActiveDownloadState building, GrabListEntry
  reconstruction, grab_most_wanted.
- lib/download_processing.py (~1230 lines) owns everything between
  "album finished downloading" and "importer queue takes over":
  materialization from event-stamped paths, beets validation, staging
  moves, auto-import dispatch, rejection handling, abandoned-auto-import
  recovery.
- lib/slskd_transfers.py (~440 lines) owns transfer-level slskd
  interaction: enqueue/status/cancel, snapshot fetch, transfer-ID
  re-derivation and matching.

lib.download.spectral_analyze is deleted outright — no callers;
lib.measurement has the copy production uses.

Consumers (cratedigger.py, lib/enqueue.py, scripts/importer.py,
scripts/import_preview_worker.py) import from the new homes. Test
imports and patch targets follow the functions: youtube-rescue and
preview-worker tests patch lib.download_processing.* (their consumer
imports from the defining module), while poll tests keep patching the
lib.download bindings the poll loop actually calls. The mock-audit
alias map gains dp_mod → lib.download_processing.
