+++ CLAUDE.md
@@ -45,7 +45,7 @@ If a future design decision drifts toward "good enough" matches, "smart" default
 - **MusicBrainz mirror** (`http://192.168.1.35:5200`) — local MB mirror. See `docs/musicbrainz-mirror.md`.
 - **Quality model** — codec-aware rank comparison (LOSSLESS > TRANSPARENT > EXCELLENT > ...). Every measurement classifies into a `QualityRank` band; gate compares against `cfg.quality_ranks.gate_min_rank` (default EXCELLENT). See `docs/quality-ranks.md` and `docs/quality-verification.md`.
 - **User cooldowns** — global 3-day cooldowns for Soulseek users with 5 consecutive failures. See `docs/cooldowns.md`.
-- **slskd client + event ingestion** (issue #146) — all slskd HTTP calls go through the in-repo typed client `lib/slskd_client.py` (the `slskd-api` PyPI dep is gone; legacy endpoints keep its dict/list return shapes, events return msgspec Structs with the double-decoded `data` payload). Phase 1 of the events refactor runs at the top of every poll cycle (`lib/slskd_events.py::ingest_download_file_events`): it pages `/api/v0/events` newest-first from a persisted single-row cursor (`slskd_event_cursor`), stamps each `DownloadFileComplete.localFilename` onto the matching `(username, remote filename)` file in `active_download_state`, and `process_completed_album` moves files from the event-stream path. **Phase 2 is active: `local_path` wins when stamped and on disk; the resolver (`resolve_slskd_local_path`) is the logged fallback** (`EVENT-PATH: chosen=resolver` at WARNING — pre-bootstrap completions and missed events). Phase 3 (delete the resolver + ticks regex + size-match, wire `cancel_and_delete` to `DownloadDirectoryComplete`, then split download.py) is a **rate judgment, not a zero-grep**: known-benign `chosen=resolver` sources are pre-bootstrap completions, the residual completion-vs-event-write race within one snapshot fetch, and `cursor_gap` truncation on very heavy days — go ahead when every fallback line traces to that residue. Phase-1 evidence for the flip: 174 match=True, 0 match=False (2026-07-02).
+- **slskd client + event ingestion** (issue #146, complete) — all slskd HTTP calls go through the in-repo typed client `lib/slskd_client.py` (the `slskd-api` PyPI dep is gone; legacy endpoints keep its dict/list return shapes, events return msgspec Structs with the double-decoded `data` payload). Event ingestion runs at the top of every poll cycle (`lib/slskd_events.py::ingest_download_file_events`): it pages `/api/v0/events` newest-first from a persisted single-row cursor (`slskd_event_cursor`) and stamps each `DownloadFileComplete.localFilename` onto the matching `(username, remote filename)` file in `active_download_state`. **The stamp is the ONLY source of completed-file locations** — the reverse-engineering resolver was deleted in phase 3 (2026-07-02). An unstamped file at materialize time is a hard failure (`EVENT-PATH MISSING` at ERROR); the poller retries within `PROCESSING_MATERIALIZE_GRACE_S` (covers the benign completion-vs-event-write race, which resolves on the next cycle's ingest) and past it self-heals the request to `wanted` for re-download. `cancel_and_delete` is likewise event-driven: it deletes files at their stamped paths (fresh `recent_completion_paths` events-page lookup for not-yet-stamped completions, consuming both `DownloadFileComplete` and `DownloadDirectoryComplete`) and prunes only empty directories — never an inferred-folder rmtree. Module layout: `lib/download.py` (poll state machine + orchestration), `lib/download_processing.py` (staging/materialization/validation dispatch), `lib/slskd_transfers.py` (enqueue/status/cancel/transfer-ID helpers), `lib/slskd_events.py` (feed ingestion).
 - **Persisted search plans** — every `wanted` request carries a deterministic plan generated by `lib.search.generate_search_plan` and persisted via `lib.search_plan_service.SearchPlanService`. Phase 2 reads `get_wanted_searchable(SEARCH_PLAN_GENERATOR_ID, ...)` and consumes plan-items by ordinal; the executor never recomputes variants. Atomic consumed-attempt writes guard against stale completions after mid-flight regeneration. Bump `SEARCH_PLAN_GENERATOR_ID` in `lib/search.py` whenever generator output changes. See `docs/persisted-search-plans-rollout.md`, `docs/pipeline-db-schema.md`.
 - **Unfindable detection** — dedicated `cratedigger-unfindable.service` oneshot on a daily systemd timer (bounded batch per run, roughly weekly per-request cadence). Categorises wanted requests into a 4-bucket taxonomy (`artist_absent`, `album_absent_artist_present`, `one_track_structural`, `wrong_pressing_available`) and writes `album_requests.unfindable_category`. **Lives in its own systemd unit, NOT inline in the 5-min `cratedigger.service` loop**, because R20 ("the system never stops searching") forbids the regular search cadence from being throttled by detection state — the structural separation makes that invariant enforceable at the systemd level. Long-tail rescues (a previously-unfindable request that finally imports) populate `rescued_at` + `prior_unfindable_category` atomically with the import. See `lib/unfindable_detection_service.py`, `docs/search-plan-iter2-deploy.md`.
 - **Triage subsystem** — `lib/triage_service.py` composes unfindable categorisation + field-resolution telemetry + search-log forensics into one typed `TriageResult: msgspec.Struct`. Operator-facing via `pipeline-cli triage show <id>` / `pipeline-cli triage list --filter=<spec>` and `GET /api/triage/<id>` / `GET /api/triage/list`. Filter syntax: `all`, `unfindable[:<category>]`, `data_quality[:<field>]`, `data_quality:status=<status>` (note `unresolved_4xx_client` lives in the **status** column, not `reason_code`), `data_quality:reason=<code>` (e.g. `http_400`), `search_not_converting`. `list_triage` is bounded to 4 DB queries at any page size (N+1 mitigation enforced by `tests/test_triage_service.py::TestListTriageN1Guard`). Read-only (`show`/`list`); replaced-row (frozen audit) inclusion is pinned by `test_list_includes_replaced_rows`. See `docs/search-plan-iter2-deploy.md`.
