+++ lib/download.py
@@ -1,69 +1,58 @@
-"""Download processing — monitoring, completion, and orchestration.
+"""Download polling — the poll state machine and search orchestration.
 
-Extracted from cratedigger.py. All functions receive a CratediggerContext
-instead of reading module-level globals.
+All functions receive a CratediggerContext instead of reading
+module-level globals. Split (issue #146 phase 3): staging /
+materialization / validation-dispatch live in
+lib/download_processing.py; slskd transfer helpers in
+lib/slskd_transfers.py; event-feed ingestion in
+lib/slskd_events.py.
 """
 
 from __future__ import annotations
 
 import logging
 import os
-import re
-import shutil
-import time
-from dataclasses import dataclass
 from datetime import datetime, timezone
 from contextlib import AbstractContextManager
-from typing import Any, Callable, Literal, Protocol, TYPE_CHECKING, runtime_checkable
+from typing import Any, Callable, Protocol, TYPE_CHECKING, runtime_checkable
 
 
+from lib.download_processing import (
+    _abandon_request_scoped_auto_import,
+    _canonical_import_folder_path,
+    _log_post_move_resume_blocked,
+    _materialize_processing_dir,
+    process_completed_album,
+)
 from lib.download_recovery import (
     classify_processing_path,
     reconcile_processing_current_path,
 )
 from lib.grab_list import GrabListEntry, DownloadFile
-from lib.processing_paths import (
-    canonical_processing_path,
-    directory_has_entries,
-    normalize_processing_path,
-    normalize_source_dirs,
-    path_is_within_root,
-    stage_to_ai_path,
-    stage_to_ai_root,
-)
+from lib.processing_paths import directory_has_entries
 from lib.quality import (ActiveDownloadState, ActiveDownloadFileState,
                          CooldownConfig,
-                         DownloadDecision, ValidationResult,
+                         DownloadDecision,
                          decide_download_action,
-                         compute_effective_override_bitrate,
-                         extract_usernames,
-                         rejection_backfill_override)
+                         extract_usernames)
 from lib import transitions
-from lib.import_dispatch import (DispatchOutcome, QualityGateFn,
-                                 _build_download_info,
-                                 _record_rejection_and_maybe_requeue,
-                                 _requeue_import_job_to_preview,
-                                 dispatch_import_core)
-from lib.import_evidence import (
-    CandidateEvidenceActionResult,
-    ensure_candidate_evidence_for_action,
-)
-from lib.import_manifest import (
-    check_audio_manifest,
-    move_failed_import_curated,
-    tracked_audio_paths_for_downloads,
-)
+from lib.import_dispatch import DispatchOutcome, _build_download_info
 from lib.import_queue import (
     IMPORT_JOB_AUTOMATION,
     ImportJob,
     automation_import_dedupe_key,
     automation_import_payload,
 )
-from lib.staged_album import StagedAlbum
-from lib.util import (
-    move_abandoned_auto_import,
-    log_validation_result,
+from lib.slskd_transfers import (
+    _all_files_remotely_queued,
+    _get_all_downloads_snapshot,
+    cancel_and_delete,
+    downloads_all_done,
+    rederive_transfer_ids,
+    slskd_do_enqueue,
+    slskd_download_status,
 )
+from lib.staged_album import StagedAlbum
 
 if TYPE_CHECKING:
     from lib.context import CratediggerContext
@@ -155,1499 +144,6 @@ MAX_FILE_RETRIES = 5
 # Generous relative to the 5-min cycle: the benign completion-vs-event
 # race resolves on the very next cycle.
 PROCESSING_MATERIALIZE_GRACE_S = 3600
-AUTO_TRIAGE_EXCLUDED_REJECTION_SCENARIOS: frozenset[str] = frozenset({
-    "audio_corrupt",
-    "spectral_reject",
-})
-ABANDONED_AUTO_IMPORT_SCENARIO = "abandoned_auto_import"
-_ABANDON_PATH_PRESENT = "present"
-_ABANDON_PATH_ABSENT = "absent"
-_ABANDON_PATH_UNKNOWN = "unknown"
-
-
-# Lazy import for spectral analysis — avoids hard dep on sox at import time
-def spectral_analyze(folder: str, trim_seconds: int = 30) -> Any:
-    """Proxy to spectral_check.analyze_album (lazy import)."""
-    from lib.spectral_check import analyze_album
-    return analyze_album(folder, trim_seconds=trim_seconds)
-
-
-def _run_post_rejection_wrong_match_cleanup(
-    ctx: "CratediggerContext",
-    download_log_id: object,
-    *,
-    scenario: str | None,
-    import_job_id: int | None = None,
-) -> Any:
-    """Evaluate newly-created Wrong Matches rows through importer cleanup.
-
-    This runs after the rejected download_log row exists and only for the
-    review-queue scenarios that Wrong Matches exposes. Bad-file scenarios have
-    their own buckets and should not be deleted through wrong-match policy.
-    """
-    if not isinstance(download_log_id, int) or isinstance(download_log_id, bool):
-        return None
-    if scenario in AUTO_TRIAGE_EXCLUDED_REJECTION_SCENARIOS:
-        return None
-    if ctx.pipeline_db_source is None:
-        return None
-    get_db = getattr(ctx.pipeline_db_source, "_get_db", None)
-    if get_db is None:
-        return None
-    try:
-        from lib.wrong_match_cleanup_service import cleanup_wrong_match
-
-        db = get_db()
-        if import_job_id is not None:
-            evidence_id = db.get_import_job_candidate_evidence_id(import_job_id)
-            if evidence_id is not None:
-                db.set_download_log_candidate_evidence(download_log_id, evidence_id)
-        result = cleanup_wrong_match(
-            db,
-            download_log_id,
-            ignore_import_job_id=import_job_id,
-        )
-        logger.info(
-            "WRONG-MATCH CLEANUP: download_log_id=%s outcome=%s verdict=%s reason=%s",
-            download_log_id,
-            getattr(result, "outcome", None),
-            getattr(result, "verdict", None),
-            getattr(result, "reason", None),
-        )
-        return result
-    except Exception:
-        logger.exception(
-            "WRONG-MATCH CLEANUP FAILED: download_log_id=%s",
-            download_log_id,
-        )
-        return None
-
-
-# === slskd file locations ===
-#
-# The authoritative local path of every completed download comes from
-# slskd's DownloadFileComplete event, stamped onto
-# ``active_download_state.files[].local_path`` by
-# ``lib.slskd_events.ingest_download_file_events`` at the top of each
-# poll cycle (issue #146). There is no on-disk path inference: a
-# completed file without a stamp is a hard failure.
-
-_REQUEST_SCOPED_STAGE_SUFFIX = re.compile(r" \[request-\d+\]$")
-
-
-def _is_request_scoped_auto_import_path(
-    *,
-    current_path: str,
-    staging_dir: str,
-) -> bool:
-    """Return True when ``current_path`` is under auto-import request staging."""
-    normalized_path = normalize_processing_path(current_path)
-    if not _REQUEST_SCOPED_STAGE_SUFFIX.search(os.path.basename(normalized_path)):
-        return False
-    return path_is_within_root(
-        normalized_path,
-        stage_to_ai_root(staging_dir=staging_dir, auto_import=True),
-    )
-
-
-def _canonical_import_folder_path(
-    album_data: GrabListEntry,
-    slskd_download_dir: str,
-) -> str:
-    return canonical_processing_path(
-        artist=album_data.artist,
-        title=album_data.title,
-        year=album_data.year,
-        slskd_download_dir=slskd_download_dir,
-    )
-
-
-def _source_dirs_for_album(album_data: GrabListEntry) -> list[str]:
-    return normalize_source_dirs(
-        [file.file_dir for file in album_data.files if file.file_dir],
-    )
-
-
-# === slskd transfer helpers ===
-
-
-@dataclass(frozen=True)
-class SlskdEnqueueOutcome:
-    """Structured result for one slskd enqueue request."""
-
-    status: Literal["accepted", "rejected", "unknown"]
-    downloads: list[DownloadFile] | None = None
-
-def cancel_and_delete(files: list[Any], ctx: CratediggerContext) -> bool:
-    """Cancel downloads and delete their completed payloads.
-
-    Deletion targets come from slskd's own answers only (issue #146
-    phase 3): the ingested ``local_path`` stamp, topped up by a fresh
-    events-page lookup for files that completed after this cycle's
-    ingest pass. Files are removed individually and their directories
-    pruned only when empty — never a recursive delete of an inferred
-    folder, which nuked unrelated albums sharing a generic leaf name
-    (the ``CD1/`` hazard). In-flight partials are slskd's to clean up.
-
-    Returns whether every requested transfer had an ID and cancel call
-    completed. Callers that only need best-effort cleanup can ignore it.
-    """
-    ok = True
-    for file in files:
-        if not file.id:
-            ok = False
-            continue  # Transfer vanished or never assigned — skip cancel
-        try:
-            if not ctx.slskd.transfers.cancel_download(
-                username=file.username,
-                id=file.id,
-            ):
-                ok = False
-        except Exception:
-            ok = False
-            logger.warning(f"Failed to cancel download {file.filename} for {file.username}",
-                           exc_info=True)
-    _delete_completed_payloads(files, ctx)
-    return ok
-
-
-def _delete_completed_payloads(
-    files: list[Any],
-    ctx: CratediggerContext,
-) -> None:
-    """Remove completed files at their authoritative local paths."""
-    from lib.slskd_events import recent_completion_paths
-
-    root = ctx.cfg.slskd_download_dir
-    needs_lookup = any(not getattr(file, "local_path", None) for file in files)
-    recent = (
-        recent_completion_paths(ctx.slskd)
-        if needs_lookup
-        else None
-    )
-
-    prune_dirs: list[str] = []
-    for file in files:
-        local_path = file.local_path or (
-            recent.files.get((file.username, file.filename))
-            if recent is not None
-            else None
-        )
-        if not local_path:
-            continue
-        if not path_is_within_root(local_path, root):
-            logger.warning(
-                "Refusing to delete %s for %s — outside the slskd "
-                "download root", local_path, file.filename)
-            continue
-        try:
-            os.remove(local_path)
-        except FileNotFoundError:
-            pass
-        except OSError:
-            logger.warning("Failed to delete %s", local_path, exc_info=True)
-            continue
-        parent = os.path.dirname(local_path)
-        if parent not in prune_dirs:
-            prune_dirs.append(parent)
-
-    if recent is not None:
-        for file in files:
-            local_dir = recent.directories.get((file.username, file.file_dir))
-            if local_dir and local_dir not in prune_dirs:
-                prune_dirs.append(local_dir)
-
-    for prune_dir in prune_dirs:
-        if not path_is_within_root(prune_dir, root):
-            continue
-        if os.path.realpath(prune_dir) == os.path.realpath(root):
-            continue
-        try:
-            os.rmdir(prune_dir)
-        except OSError:
-            pass  # Non-empty (shared with another album) or already gone.
-
-
-def slskd_download_status(downloads: list[Any], ctx: CratediggerContext,
-                          *, snapshot: list[dict[str, Any]] | None = None) -> bool:
-    """Get status of each download file from slskd API.
-
-    When snapshot is provided, matches locally against the pre-fetched bulk
-    download list instead of making per-file API calls.
-    """
-    ok = True
-    for file in downloads:
-        try:
-            if snapshot is not None:
-                transfer = match_transfer(snapshot, file.filename, username=file.username)
-                if transfer is not None:
-                    file.status = dict(transfer)
-                else:
-                    file.status = None
-                    ok = False
-            else:
-                status = ctx.slskd.transfers.get_download(file.username, file.id)
-                file.status = status
-        except Exception:
-            logger.exception(f"Error getting download status of {file.filename}")
-            file.status = None
-            ok = False
-    return ok
-
-
-def _is_user_offline_http_error(exc: BaseException) -> bool:
-    """slskd surfaces a peer-offline rejection as an HTTPError whose
-    response body contains 'appears to be offline' (verified against
-    Soulseek.UserOfflineException — see the 2026-05-08 pooyork incident).
-
-    Match on body substring rather than status code: slskd has shipped 400 /
-    500 / 504 across versions for this case, but the body string is durable.
-    Detection is structural — any exception carrying a ``.response`` with
-    a readable ``.text`` containing the marker counts. Avoiding
-    ``isinstance(requests.exceptions.HTTPError)`` keeps the helper safe
-    when ``sys.modules["requests"]`` is monkey-patched in tests.
-    """
-    response = getattr(exc, "response", None)
-    if response is None:
-        return False
-    try:
-        body = response.text or ""
-    except Exception:
-        return False
-    if not isinstance(body, str):
-        return False
-    return "appears to be offline" in body.lower()
-
-
-def slskd_enqueue_with_outcome(
-    username: str,
-    files: list[dict[str, Any]],
-    file_dir: str,
-    ctx: CratediggerContext,
-) -> SlskdEnqueueOutcome:
-    """Enqueue files for download via slskd with an explicit outcome."""
-    try:
-        enqueue = ctx.slskd.transfers.enqueue(username=username, files=files)
-    except Exception as exc:
-        if _is_user_offline_http_error(exc):
-            logger.info(
-                "slskd reports peer %s offline at enqueue; classifying as rejected",
-                username,
-            )
-            return SlskdEnqueueOutcome(status="rejected")
-        logger.debug("Enqueue failed", exc_info=True)
-        return SlskdEnqueueOutcome(status="unknown")
-    if not enqueue:
-        return SlskdEnqueueOutcome(status="rejected")
-
-    # Poll for transfer IDs — slskd needs time to register the enqueue.
-    # Typically resolves in 1-2s; max 10s before giving up.
-    max_wait = 10.0
-    interval = 1.0
-    elapsed = 0.0
-    download_list: list[dict[str, Any]] | None = None
-
-    while elapsed < max_wait:
-        time.sleep(interval)
-        elapsed += interval
-        download_list = _get_all_downloads_snapshot(
-            ctx.slskd,
-            purpose=f"download status for {username} after enqueue",
-        )
-        if download_list is None:
-            continue
-        if all(
-            match_transfer_id(download_list, f["filename"], username=username)
-            is not None
-            for f in files
-        ):
-            break
-
-    downloads: list[DownloadFile] = []
-    for file in files:
-        transfer_id = (
-            match_transfer_id(
-                download_list,
-                file["filename"],
-                username=username,
-            )
-            if download_list is not None
-            else None
-        )
-        downloads.append(DownloadFile(
-            filename=file["filename"],
-            id=transfer_id or "",
-            file_dir=file_dir,
-            username=username,
-            size=file["size"],
-        ))
-
-    reconciled = sum(1 for download in downloads if download.id)
-    if reconciled != len(downloads):
-        logger.warning(
-            "slskd accepted enqueue for %s but only reconciled %s/%s "
-            "transfer IDs; tracking filenames for next-cycle rederivation",
-            username,
-            reconciled,
-            len(downloads),
-        )
-    return SlskdEnqueueOutcome(status="accepted", downloads=downloads)
-
-
-def slskd_do_enqueue(username: str, files: list[dict[str, Any]],
-                     file_dir: str, ctx: CratediggerContext) -> list[DownloadFile] | None:
-    """Enqueue files for download via slskd. Returns DownloadFile list or None."""
-    outcome = slskd_enqueue_with_outcome(username, files, file_dir, ctx)
-    if outcome.status != "accepted":
-        return None
-    return outcome.downloads
-
-
-def downloads_all_done(downloads: list[Any]) -> tuple[bool, list[Any] | None, int]:
-    """Check status of all files. Returns (all_done, error_list_or_none, remote_queue_count)."""
-    all_done = True
-    error_list: list[Any] = []
-    remote_queue = 0
-    for file in downloads:
-        if file.status is not None:
-            state = file.status.get("state", "")
-            if state != "Completed, Succeeded":
-                all_done = False
-            if state in (
-                "Completed, Cancelled",
-                "Completed, TimedOut",
-                "Completed, Errored",
-                "Completed, Rejected",
-                "Completed, Aborted",
-            ):
-                error_list.append(file)
-            if file.status["state"] == "Queued, Remotely":
-                remote_queue += 1
-    return all_done, error_list if error_list else None, remote_queue
-
-
-def _all_files_remotely_queued(downloads: list[Any], remote_queue_count: int) -> bool:
-    """Return True when every tracked file is still queued on the remote peer.
-
-    This is intentionally separate from stalled transfer detection: a file that has
-    never started uploading should be governed by the remote queue timeout, not the
-    no-progress timeout used for active transfers.
-    """
-    return bool(downloads) and remote_queue_count == len(downloads)
-
-
-
-
-# === Download completion processing ===
-def _log_post_move_resume_blocked(
-    album_data: GrabListEntry,
-    *,
-    current_path: str,
-    detail: str,
-) -> None:
-    logger.error(
-        "POST-MOVE RESUME BLOCKED: request_id=%s %s - %s "
-        "current_path=%s %s See docs/advisory-locks.md.",
-        album_data.db_request_id,
-        album_data.artist,
-        album_data.title,
-        current_path,
-        detail,
-    )
-
-
-def _request_import_subprocess_started(
-    db: DownloadDB | None,
-    request_id: int | None,
-) -> bool | None:
-    """Return subprocess-start evidence, or None when ownership is unknown."""
-    if request_id is None or db is None:
-        return None
-    try:
-        row = db.get_request(request_id)
-    except Exception:
-        logger.debug(
-            "Failed to read active_download_state for resume guard",
-            exc_info=True,
-        )
-        return None
-    if not row:
-        return None
-    raw_state = row.get("active_download_state")
-    if not raw_state:
-        return False
-    try:
-        state = (
-            ActiveDownloadState.from_dict(raw_state)
-            if isinstance(raw_state, dict)
-            else ActiveDownloadState.from_json(str(raw_state))
-        )
-    except Exception:
-        logger.debug(
-            "Failed to parse active_download_state for resume guard",
-            exc_info=True,
-        )
-        return None
-    return state.import_subprocess_started_at is not None
-
-
-def _import_subprocess_already_started(
-    db: DownloadDB | None,
-    request_id: int | None,
-) -> bool:
-    """Did a previous attempt actually launch ``import_one.py`` for this row?
-
-    The auto-import resume guard blocks retries when files live at the
-    request-scoped staged path because a prior subprocess may have
-    started writing to beets. That guard is correct only when a
-    subprocess actually started — files-at-staged is a necessary but not
-    sufficient signal. The 2026-05-04 wedge accumulated 5788 failed
-    importer jobs because the guard fired even when the subprocess had
-    never launched (crash window between staged-move and subprocess
-    spawn). See ``docs/advisory-locks.md``.
-
-    With the ``import_subprocess_started_at`` flag set in
-    ``ActiveDownloadState`` immediately before ``run_import_one(...)``,
-    this helper returns ``True`` only when the flag is set: the
-    necessary AND sufficient evidence that the subprocess could have
-    written to beets.
-
-    Returns ``True`` (block) if state is unreachable — fail safe; the
-    operator can still recover manually. Returns ``False`` (permit
-    retry) only on positive evidence the subprocess never launched.
-    """
-    return _request_import_subprocess_started(db, request_id) is not False
-
-
-def _probe_abandon_path_liveness(path: str) -> str:
-    """Return whether a staged path is definitely present, absent, or unknown."""
-    try:
-        os.stat(path)
-    except FileNotFoundError:
-        return _ABANDON_PATH_ABSENT
-    except OSError:
-        logger.exception(
-            "ABANDON AUTO-IMPORT BLOCKED: could not stat current_path=%s",
-            path,
-        )
-        return _ABANDON_PATH_UNKNOWN
-    return _ABANDON_PATH_PRESENT
-
-
-def _restore_abandoned_auto_import(
-    *,
-    failed_path: str | None,
-    current_path: str,
-) -> None:
-    if failed_path is None:
-        return
-    try:
-        if os.path.exists(failed_path) and not os.path.exists(current_path):
-            os.makedirs(os.path.dirname(current_path), exist_ok=True)
-            shutil.move(failed_path, current_path)
-    except Exception:
-        logger.exception(
-            "ABANDON AUTO-IMPORT ROLLBACK FAILED: failed_path=%s current_path=%s",
-            failed_path,
-            current_path,
-        )
-
-
-def _commit_abandoned_auto_import(
-    db: DownloadDB,
-    *,
-    request_id: int,
-    current_path: str,
-    dl_info: Any,
-    detail: str,
-    validation_result: str | None,
-) -> bool:
-    log_id = db.abandon_auto_import_request(
-        request_id=request_id,
-        current_path=current_path,
-        soulseek_username=dl_info.username,
-        filetype=dl_info.filetype,
-        beets_scenario=ABANDONED_AUTO_IMPORT_SCENARIO,
-        beets_detail=detail,
-        outcome="failed",
-        staged_path=current_path,
-        error_message=detail,
-        validation_result=validation_result,
-    )
-    return log_id is not None
-
-
-def _abandon_interrupted_auto_import(
-    album_data: GrabListEntry,
-    *,
-    request_id: int,
-    current_path: str,
-    db: DownloadDB,
-    detail: str,
-) -> bool:
-    """Quarantine an interrupted auto-import attempt and redownload later."""
-    path_state = _probe_abandon_path_liveness(current_path)
-    if path_state == _ABANDON_PATH_UNKNOWN:
-        return False
-
-    failed_path: str | None = None
-    if path_state == _ABANDON_PATH_PRESENT:
-        try:
-            failed_path = move_abandoned_auto_import(current_path)
-        except Exception:
-            logger.exception(
-                "ABANDON AUTO-IMPORT FAILED: request_id=%s current_path=%s",
-                request_id,
-                current_path,
-            )
-            return False
-
-    dl_info = _build_download_info(album_data)
-    validation_result: str | None = None
-    if failed_path is not None:
-        validation_result = ValidationResult(
-            valid=False,
-            scenario=ABANDONED_AUTO_IMPORT_SCENARIO,
-            detail=detail,
-            path=current_path,
-            soulseek_username=dl_info.username,
-            download_folder=current_path,
-            failed_path=failed_path,
-        ).to_json()
-
-    logger.warning(
-        "ABANDON AUTO-IMPORT: request_id=%s %s - %s current_path=%s "
-        "failed_path=%s detail=%s",
-        request_id,
-        album_data.artist,
-        album_data.title,
-        current_path,
-        failed_path,
-        detail,
-    )
-    try:
-        committed = _commit_abandoned_auto_import(
-            db,
-            request_id=request_id,
-            current_path=current_path,
-            dl_info=dl_info,
-            detail=detail,
-            validation_result=validation_result,
-        )
-    except Exception:
-        _restore_abandoned_auto_import(
-            failed_path=failed_path,
-            current_path=current_path,
-        )
-        logger.exception(
-            "ABANDON AUTO-IMPORT DB COMMIT FAILED: request_id=%s current_path=%s",
-            request_id,
-            current_path,
-        )
-        return False
-    if not committed:
-        _restore_abandoned_auto_import(
-            failed_path=failed_path,
-            current_path=current_path,
-        )
-        logger.warning(
-            "ABANDON AUTO-IMPORT SKIPPED: request_id=%s current_path=%s "
-            "row ownership changed before commit",
-            request_id,
-            current_path,
-        )
-        return False
-    return True
-
-
-def _abandon_request_scoped_auto_import(
-    album_data: GrabListEntry,
-    *,
-    request_id: int | None,
-    current_path: str,
-    current_path_kind: str,
-    db: DownloadDB | None,
-    detail: str,
-) -> bool:
-    if (
-        request_id is None
-        or db is None
-        or current_path_kind != "request_scoped_auto_import_staged"
-    ):
-        return False
-    if not album_data.mb_release_id:
-        _log_post_move_resume_blocked(
-            album_data,
-            current_path=current_path,
-            detail=(
-                "already lives at the request-scoped auto-import staged "
-                "path but has no release id for the liveness lock; "
-                "manual recovery is required."
-            ),
-        )
-        return False
-    try:
-        from lib.pipeline_db import (
-            ADVISORY_LOCK_NAMESPACE_RELEASE,
-            release_id_to_lock_key,
-        )
-
-        with db.advisory_lock(
-            ADVISORY_LOCK_NAMESPACE_RELEASE,
-            release_id_to_lock_key(album_data.mb_release_id),
-        ) as acquired:
-            if not acquired:
-                _log_post_move_resume_blocked(
-                    album_data,
-                    current_path=current_path,
-                    detail=(
-                        "already lives at the request-scoped auto-import "
-                        "staged path, but the release import lock is held; "
-                        "leaving it for the active importer."
-                    ),
-                )
-                return False
-            return _abandon_interrupted_auto_import(
-                album_data,
-                request_id=request_id,
-                current_path=current_path,
-                db=db,
-                detail=detail,
-            )
-    except Exception:
-        logger.exception(
-            "ABANDON AUTO-IMPORT LOCK CHECK FAILED: request_id=%s current_path=%s",
-            request_id,
-            current_path,
-        )
-        return False
-
-
-def _materialize_processing_dir(
-    album_data: GrabListEntry,
-    staged_album: StagedAlbum,
-    ctx: CratediggerContext,
-) -> bool | None:
-    """Ensure ``staged_album.current_path`` holds the album's local files."""
-    canonical_path = _canonical_import_folder_path(
-        album_data, ctx.cfg.slskd_download_dir)
-    db = (ctx.pipeline_db_source._get_db()
-          if ctx.pipeline_db_source is not None else None)
-    request_id = album_data.db_request_id
-    if request_id is None and _is_request_scoped_auto_import_path(
-        current_path=staged_album.current_path,
-        staging_dir=ctx.cfg.beets_staging_dir,
-    ):
-        _log_post_move_resume_blocked(
-            album_data,
-            current_path=staged_album.current_path,
-            detail=(
-                "already lives at the request-scoped auto-import staged "
-                "path but is missing db_request_id. Automatic retry is "
-                "disabled because import ownership can no longer be "
-                "verified; manual recovery is required."
-            ),
-        )
-        return None
-    current_path_location = classify_processing_path(
-        current_path=staged_album.current_path,
-        artist=album_data.artist,
-        title=album_data.title,
-        year=album_data.year,
-        request_id=request_id or 0,
-        staging_dir=ctx.cfg.beets_staging_dir,
-        slskd_download_dir=ctx.cfg.slskd_download_dir,
-    )
-
-    if current_path_location.kind != "canonical":
-        subprocess_started = _request_import_subprocess_started(
-            db, request_id)
-        if current_path_location.kind == "request_scoped_auto_import_staged":
-            if subprocess_started is True:
-                handled = _abandon_request_scoped_auto_import(
-                    album_data,
-                    request_id=request_id,
-                    current_path=staged_album.current_path,
-                    current_path_kind=current_path_location.kind,
-                    db=db,
-                    detail=(
-                        "Abandoned interrupted auto-import; queued for "
-                        "redownload"
-                    ),
-                )
-                return False if handled else None
-            if subprocess_started is None:
-                _log_post_move_resume_blocked(
-                    album_data,
-                    current_path=staged_album.current_path,
-                    detail=(
-                        "already lives at the request-scoped auto-import "
-                        "staged path but import ownership could not be "
-                        "verified; manual recovery is required."
-                    ),
-                )
-                return None
-        if not os.path.isdir(staged_album.current_path):
-            if (
-                current_path_location.blocks_post_move_retry
-                and subprocess_started is not False
-            ):
-                _log_post_move_resume_blocked(
-                    album_data,
-                    current_path=staged_album.current_path,
-                    detail=(
-                        "already lives at the request-scoped auto-import "
-                        "staged path but the directory is missing. "
-                        "Automatic retry is disabled because beets may "
-                        "already have consumed the staged folder; manual "
-                        "recovery is required."
-                    ),
-                )
-                return None
-            logger.error(f"Current staged path missing: {staged_album.current_path}")
-            return False
-        staged_album.bind_import_paths(album_data.files)
-        missing_paths: list[str] = []
-        for file in album_data.files:
-            import_path = file.import_path
-            assert import_path is not None
-            if not os.path.isfile(import_path):
-                missing_paths.append(import_path)
-        if missing_paths:
-            if (
-                current_path_location.blocks_post_move_retry
-                and subprocess_started is not False
-            ):
-                _log_post_move_resume_blocked(
-                    album_data,
-                    current_path=staged_album.current_path,
-                    detail=(
-                        "already lives at the request-scoped auto-import "
-                        f"staged path but tracked files are missing ({', '.join(missing_paths)}). "
-                        "Automatic retry is disabled because import may "
-                        "already have started; manual recovery is required."
-                    ),
-                )
-                return None
-            logger.error(
-                "Current staged path is missing tracked files: %s",
-                ", ".join(missing_paths),
-            )
-            return False
-        if (
-            current_path_location.blocks_auto_import_dispatch
-            and subprocess_started is not False
-        ):
-            detail = (
-                "already lives at the request-scoped auto-import staged "
-                "path. Automatic retry is disabled to avoid duplicate "
-                "import; manual recovery is required."
-            )
-            if current_path_location.kind == "legacy_shared_staged":
-                detail = (
-                    "already lives at the legacy shared staged path. "
-                    "Automatic retry is disabled because the path is "
-                    "ambiguous across editions; manual recovery is required."
-                )
-            _log_post_move_resume_blocked(
-                album_data,
-                current_path=staged_album.current_path,
-                detail=detail,
-            )
-            return None
-        album_data.import_folder = staged_album.current_path
-        return True
-
-    # Pre-flight: every file must carry a stamped, on-disk local_path from
-    # slskd's DownloadFileComplete event — or already-moved evidence at the
-    # destination. Checked for the whole album BEFORE any move so a
-    # missing stamp never causes move-then-rollback churn. A stamp can be
-    # legitimately absent for one cycle (completion-vs-event-write race);
-    # the poller retries within ``PROCESSING_MATERIALIZE_GRACE_S`` and
-    # self-heals to re-download past it.
-    missing_stamps: list[str] = []
-    for file in album_data.files:
-        dst_file = staged_album.import_path_for(file)
-        file.import_path = dst_file
-        src = file.local_path
-        if src is not None and os.path.exists(src):
-            continue
-        if os.path.exists(dst_file):
-            continue  # Already moved by a prior crashed attempt.
-        if src is None:
-            missing_stamps.append(f"{file.filename} (not_stamped)")
-        else:
-            missing_stamps.append(f"{file.filename} (stale_stamp: {src})")
-    if missing_stamps:
-        logger.error(
-            "EVENT-PATH MISSING: request_id=%s %s - %s has no authoritative "
-            "local path for %d file(s): %s. The DownloadFileComplete event "
-            "was never ingested (pre-bootstrap completion or cursor gap) or "
-            "the stamped file vanished from disk.",
-            album_data.db_request_id,
-            album_data.artist,
-            album_data.title,
-            len(missing_stamps),
-            "; ".join(missing_stamps),
-        )
-        return False
-
-    rm_dirs: list[str] = []
-    moved_files_history: list[tuple[str, str]] = []
-    if os.path.exists(canonical_path):
-        logger.info(f"Staging folder {canonical_path} already exists — "
-                    f"resuming or reusing prior attempt")
-    else:
-        try:
-            os.makedirs(canonical_path, exist_ok=True)
-        except OSError:
-            # ENAMETOOLONG, EACCES, ENOSPC, etc. Letting this propagate
-            # would abort the entire poll loop and starve every later row.
-            logger.exception(
-                "Failed to create canonical staging dir %s for request %s — "
-                "leaving for next poll cycle",
-                canonical_path,
-                album_data.db_request_id,
-            )
-            return False
-
-    for file in album_data.files:
-        dst_file = file.import_path
-        assert dst_file is not None
-        src_file = file.local_path
-        if src_file is None or not os.path.exists(src_file):
-            # Pre-flight proved the destination copy exists.
-            logger.info(f"Already-moved file detected: {dst_file} (src gone, skipping)")
-            continue
-        src_folder = os.path.dirname(src_file)
-        if src_folder not in rm_dirs:
-            rm_dirs.append(src_folder)
-        try:
-            # Destination keeps the clean remote basename (via
-            # ``import_path_for``) even when slskd appended a ``_<ticks>``
-            # collision suffix to the source.
-            shutil.move(src_file, dst_file)
-            moved_files_history.append((src_file, dst_file))
-        except Exception:
-            logger.exception(f"Failed to move: {file.filename} to temp location for import. Rolling back...")
-            for src, dst in reversed(moved_files_history):
-                try:
-                    shutil.move(dst, src)
-                except Exception:
-                    logger.exception(f"Critical failure during rollback: could not move {dst} back to {src}")
-            try:
-                os.rmdir(canonical_path)
-            except OSError:
-                logger.warning(f"Could not remove temp import directory {canonical_path}")
-            return False
-
-    for rm_dir in rm_dirs:
-        if os.path.abspath(rm_dir) == os.path.abspath(canonical_path):
-            continue
-        try:
-            os.rmdir(rm_dir)
-        except OSError:
-            logger.warning(f"Skipping removal of {rm_dir} because it's not empty.")
-
-    album_data.import_folder = staged_album.current_path
-    staged_album.persist_current_path(db)
-    return True
-
-
-def _check_staged_audio_manifest(
-    album_data: GrabListEntry,
-    staged_album: StagedAlbum,
-) -> tuple[bool, str]:
-    check = check_audio_manifest(
-        staged_album.current_path,
-        tracked_audio_paths_for_downloads(album_data.files),
-    )
-    if check.ok:
-        return True, ""
-    detail = (
-        "Staged import folder does not match the selected audio manifest: "
-        f"{check.detail()}"
-    )
-    logger.error(
-        "IMPORT MANIFEST REJECTED: request_id=%s path=%s %s",
-        album_data.db_request_id,
-        staged_album.current_path,
-        detail,
-    )
-    return False, detail
-
-
-def process_completed_album(
-    album_data: GrabListEntry,
-    failed_grab: list[Any],
-    ctx: CratediggerContext,
-    *,
-    import_job_id: int,
-    validate_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-    handle_valid_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-    dispatch_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-) -> "bool | DispatchOutcome | None":
-    """Process a fully-downloaded album: move files, tag, validate, stage/import.
-
-    Returns the local processing result:
-    - ``True`` — local non-dispatch processing succeeded. Outer caller may
-      finalize only if the request row is still ``downloading``.
-    - ``False`` — local non-dispatch processing failed. Outer caller resets
-      to ``wanted`` only if the request row is still ``downloading``.
-    - ``DispatchOutcome`` — the validation / dispatch path already owned the
-      request transition and produced an import summary for the queue owner.
-    - ``None`` — the validation / dispatch path intentionally left state
-      untouched for retry / manual recovery. Outer caller must NOT touch status.
-    """
-    staged_album = StagedAlbum.from_entry(
-        album_data,
-        default_path=_canonical_import_folder_path(
-            album_data, ctx.cfg.slskd_download_dir),
-    )
-    materialized = _materialize_processing_dir(album_data, staged_album, ctx)
-    if materialized is not True:
-        return materialized
-
-    logger.info(f"Processing completed download: {album_data.artist} - {album_data.title}")
-    if ctx.cfg.beets_validation_enabled and album_data.mb_release_id:
-        _validate = validate_fn if validate_fn is not None else _process_beets_validation
-        outcome = _validate(
-            album_data,
-            staged_album,
-            ctx,
-            import_job_id=import_job_id,
-            handle_valid_fn=handle_valid_fn,
-            dispatch_fn=dispatch_fn,
-        )
-        if outcome is not None:
-            if outcome.deferred:
-                # Release-lock contention. Propagate ``None`` so
-                # ``_run_completed_processing`` leaves the request's
-                # status, active_download_state, and staged files
-                # untouched for the next cycle to retry.
-                return None
-            # DispatchOutcome is an import summary only. Return it so the
-            # importer queue can record the real terminal job outcome, but do
-            # not let it drive fallback request-status transitions below.
-            return outcome
-    return True
-
-
-def _process_beets_validation(
-    album_data: GrabListEntry,
-    staged_album: StagedAlbum,
-    ctx: CratediggerContext,
-    *,
-    import_job_id: int,
-    handle_valid_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-    dispatch_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-) -> "DispatchOutcome | None":
-    """Beets validation sub-path of process_completed_album.
-
-    After beets validation passes, ``ensure_candidate_evidence_for_action``
-    confirms the preview worker has persisted candidate evidence keyed to
-    this import_job. Missing evidence requeues the job to preview rather
-    than measuring inline — the importer never measures, the preview
-    worker owns evidence production, and the full pipeline decider
-    (``full_pipeline_decision_from_evidence``) runs downstream of evidence.
-
-    Returns the dispatch outcome when the auto-import path fires,
-    ``None`` when beets validation rejects (``_handle_rejected_result``
-    already handles the state transition) or when the non-auto
-    redownload path takes over in ``_handle_valid_result``. Guarded
-    ownership-less rejects also return a deferred outcome so callers
-    keep the row untouched for manual recovery.
-    """
-    from lib.beets import beets_validate as _bv
-    current_path = staged_album.current_path
-    manifest_ok, manifest_detail = _check_staged_audio_manifest(
-        album_data,
-        staged_album,
-    )
-    if not manifest_ok:
-        return _reject_request_auto_import(
-            album_data,
-            ValidationResult(
-                valid=False,
-                scenario="untracked_audio",
-                detail=manifest_detail,
-                error=manifest_detail,
-                path=current_path,
-            ),
-            staged_album,
-            ctx,
-            detail=manifest_detail,
-            scenario="untracked_audio",
-            error=manifest_detail,
-        )
-    bv_result = _bv(ctx.cfg.beets_harness_path, current_path,
-                    album_data.mb_release_id, ctx.cfg.beets_distance_threshold)
-    usernames_pre = set(f.username for f in album_data.files if f.username)
-    bv_result.soulseek_username = ", ".join(sorted(usernames_pre)) if usernames_pre else None
-    bv_result.download_folder = current_path
-    bv_result.source_dirs = _source_dirs_for_album(album_data)
-    if bv_result.valid:
-        db = ctx.pipeline_db_source._get_db()
-        candidate_result = ensure_candidate_evidence_for_action(
-            db,
-            source_path=current_path,
-            import_job_id=import_job_id,
-        )
-        if not candidate_result.available:
-            reason = (
-                candidate_result.provenance.fallback_reason
-                or candidate_result.provenance.candidate_status
-                or "missing"
-            )
-            # Preview owns candidate-evidence production; the importer
-            # never measures. Requeue rather than fail; the dispatch-side
-            # requeue keeps the advisory-lock atomicity intact.
-            return _requeue_import_job_to_preview(
-                db,
-                import_job_id=import_job_id,
-                reason=reason,
-            )
-        _handle_valid = (
-            handle_valid_fn if handle_valid_fn is not None else _handle_valid_result
-        )
-        return _handle_valid(
-            album_data,
-            bv_result,
-            staged_album,
-            ctx,
-            import_job_id=import_job_id,
-            prevalidated_candidate_result=candidate_result,
-            dispatch_fn=dispatch_fn,
-        )
-    return _handle_rejected_result(
-        album_data,
-        bv_result,
-        staged_album,
-        ctx,
-        import_job_id=import_job_id,
-    )
-
-
-def _resolved_request_rejection_id(
-    album_data: GrabListEntry,
-    ctx: CratediggerContext,
-) -> tuple[Any | None, int | None]:
-    """Resolve the backing request row for defensive auto-import rejects."""
-    if ctx.pipeline_db_source is None:
-        return None, None
-    db = ctx.pipeline_db_source._get_db()
-    if album_data.db_request_id is not None:
-        return db, album_data.db_request_id
-
-    candidate_request_id = album_data.album_id
-    if not isinstance(candidate_request_id, int) or isinstance(candidate_request_id, bool):
-        return db, None
-    # ``AlbumRecord.id`` is negative on the search path, so only positive
-    # ids can safely be treated as ``album_requests.id`` candidates here.
-    if candidate_request_id <= 0:
-        return db, None
-
-    request_row = db.get_request(candidate_request_id)
-    if not isinstance(request_row, dict):
-        return db, None
-    if str(request_row.get("artist_name") or "") != album_data.artist:
-        return db, None
-    if str(request_row.get("album_title") or "") != album_data.title:
-        return db, None
-    request_year = request_row.get("year")
-    if (
-        album_data.year
-        and request_year not in (None, "")
-        and str(request_year) != album_data.year
-    ):
-        return db, None
-    album_release_id = str(album_data.mb_release_id or "")
-    request_release_id = str(request_row.get("mb_release_id") or "")
-    if bool(album_release_id) != bool(request_release_id):
-        return db, None
-    if album_release_id and request_release_id != album_release_id:
-        return db, None
-    return db, candidate_request_id
-
-
-def _reject_request_auto_import(
-    album_data: GrabListEntry,
-    bv_result: ValidationResult,
-    staged_album: StagedAlbum,
-    ctx: CratediggerContext,
-    *,
-    detail: str,
-    scenario: str,
-    error: str,
-) -> DispatchOutcome:
-    """Reject a request auto-import when ownership can be proven safely."""
-    db, request_id = _resolved_request_rejection_id(album_data, ctx)
-    if db is None or request_id is None:
-        logger.error(
-            "AUTO-IMPORT REJECT BLOCKED WITHOUT REQUEST AUDIT: album_id=%s %s - %s "
-            "(scenario=%s) could not resolve a safe pipeline request row; "
-            "files remain at %s and automatic retry/import is disabled until "
-            "manual recovery.",
-            album_data.album_id,
-            album_data.artist,
-            album_data.title,
-            scenario,
-            staged_album.current_path,
-        )
-        return DispatchOutcome(
-            success=False,
-            message=detail,
-            deferred=True,
-        )
-
-    failed_result = ValidationResult(
-        distance=bv_result.distance if bv_result.distance is not None else 0.0,
-        scenario=scenario,
-        detail=detail,
-        error=error,
-    )
-    failed_result.source_dirs = _source_dirs_for_album(album_data)
-    failed_result.failed_path = move_failed_import_curated(
-        staged_album.current_path,
-        allowed_audio=tracked_audio_paths_for_downloads(album_data.files),
-        scenario=failed_result.scenario,
-    )
-    logger.error(
-        "AUTO-IMPORT REJECTED: %s - %s — %s",
-        album_data.artist,
-        album_data.title,
-        detail,
-    )
-    log_validation_result(album_data, failed_result, ctx.cfg)
-
-    dl_info = _build_download_info(album_data)
-    if album_data.download_spectral is not None:
-        dl_info.download_spectral = album_data.download_spectral
-        dl_info.current_spectral = album_data.current_spectral
-        dl_info.existing_min_bitrate = album_data.current_min_bitrate
-        dl_info.slskd_filetype = dl_info.filetype
-        dl_info.actual_filetype = dl_info.filetype
-    download_log_id = _record_rejection_and_maybe_requeue(
-        db,
-        request_id,
-        dl_info,
-        distance=failed_result.distance if failed_result.distance is not None else 0.0,
-        scenario=failed_result.scenario or scenario,
-        detail=detail,
-        error=failed_result.error,
-        requeue=True,
-        validation_result=failed_result.to_json(),
-    )
-    _run_post_rejection_wrong_match_cleanup(
-        ctx,
-        download_log_id,
-        scenario=failed_result.scenario,
-    )
-
-    return DispatchOutcome(success=False, message=detail)
-
-
-def _handle_valid_result(
-    album_data: GrabListEntry,
-    bv_result: ValidationResult,
-    staged_album: StagedAlbum,
-    ctx: CratediggerContext,
-    *,
-    import_job_id: int | None = None,
-    prevalidated_candidate_result: CandidateEvidenceActionResult | None = None,
-    quality_gate_fn: QualityGateFn | None = None,
-    dispatch_fn: "Callable[..., DispatchOutcome | None] | None" = None,
-) -> "DispatchOutcome | None":
-    """Handle a valid beets validation result: stage and optionally auto-import.
-
-    Returns the ``DispatchOutcome`` summary from ``dispatch_import_core``
-    when the auto-import path fires (source='request', distance within
-    threshold), or ``None`` for the redownload path that just stages
-    and marks done. ``process_completed_album()`` propagates the summary
-    upward for the importer queue, but request-state changes remain owned
-    by the dispatch/finalization seam itself.
-
-    This function acquires the RELEASE advisory lock outer for the
-    auto-import path *before* ``StagedAlbum.move_to`` runs, so
-    contention is a true no-op: files stay at their current local
-    processing path, ``active_download_state.current_path`` stays
-    unchanged, and the next cycle can idempotently re-enter without
-    any extra filesystem churn. Redownload paths don't take the lock
-    — they just move into staging and mark done, so no cross-process
-    race applies.
-
-    See ``docs/advisory-locks.md`` for namespaces, keys, ordering,
-    and contention behaviour (including the staged-move rationale for
-    acquiring at this level rather than inside
-    ``dispatch_import_core``).
-    """
-    from contextlib import nullcontext
-    from lib.pipeline_db import (ADVISORY_LOCK_NAMESPACE_RELEASE,
-                                 release_id_to_lock_key)
-
-    source_type = album_data.db_source or "redownload"
-    request_id = album_data.db_request_id
-    dist = bv_result.distance if bv_result.distance is not None else 1.0
-    wants_auto_import = (
-        source_type == "request"
-        and dist <= ctx.cfg.beets_distance_threshold)
-
-    if wants_auto_import and request_id is None:
-        return _reject_request_auto_import(
-            album_data,
-            bv_result,
-            staged_album,
-            ctx,
-            detail=(
-                "Request auto-import is missing db_request_id; automatic "
-                "resume/import is disabled."
-            ),
-            scenario="request_missing_request_id",
-            error="missing_request_id",
-        )
-
-    current_path_location = classify_processing_path(
-        current_path=staged_album.current_path,
-        artist=album_data.artist,
-        title=album_data.title,
-        year=album_data.year,
-        request_id=request_id or 0,
-        staging_dir=ctx.cfg.beets_staging_dir,
-        slskd_download_dir=ctx.cfg.slskd_download_dir,
-    )
-
-    if wants_auto_import and not album_data.mb_release_id:
-        return _reject_request_auto_import(
-            album_data,
-            bv_result,
-            staged_album,
-            ctx,
-            detail="Request auto-import requires a MusicBrainz release ID",
-            scenario="request_missing_mbid",
-            error="missing_mbid",
-        )
-
-    will_auto_import = wants_auto_import
-    pdb = None
-
-    if (
-        will_auto_import
-        and current_path_location.blocks_auto_import_dispatch
-        and _import_subprocess_already_started(
-            ctx.pipeline_db_source._get_db()
-            if ctx.pipeline_db_source is not None
-            else None,
-            request_id,
-        )
-    ):
-        _log_post_move_resume_blocked(
-            album_data,
-            current_path=staged_album.current_path,
-            detail=(
-                f"already lives at the {current_path_location.display_name}. "
-                "Automatic retry is disabled to avoid duplicate import; "
-                "manual recovery is required."
-            ),
-        )
-        return DispatchOutcome(
-            success=False,
-            message=(
-                "Auto-import may already have started for this staged "
-                f"album ({album_data.mb_release_id})"
-            ),
-            deferred=True,
-        )
-
-    if will_auto_import and album_data.mb_release_id:
-        pdb = ctx.pipeline_db_source._get_db()
-        lock_ctx = pdb.advisory_lock(
-            ADVISORY_LOCK_NAMESPACE_RELEASE,
-            release_id_to_lock_key(album_data.mb_release_id))
-    else:
-        lock_ctx = nullcontext(True)
-
-    with lock_ctx as got_release_lock:
-        if not got_release_lock:
-            logger.warning(
-                f"AUTO-IMPORT DEFERRED: {album_data.artist} - "
-                f"{album_data.title} — release lock held by another "
-                f"process (mbid={album_data.mb_release_id}); skipping "
-                "staged move and dispatch. Files stay at "
-                f"{staged_album.current_path} so the next cycle can "
-                "idempotently resume from process_completed_album.")
-            return DispatchOutcome(
-                success=False,
-                message=("Another import is already in progress for "
-                         f"this release ({album_data.mb_release_id})"),
-                deferred=True,
-            )
-
-        db = (ctx.pipeline_db_source._get_db()
-              if ctx.pipeline_db_source is not None else None)
-        dest = staged_album.move_to(
-            stage_to_ai_path(
-                artist=album_data.artist,
-                title=album_data.title,
-                staging_dir=ctx.cfg.beets_staging_dir,
-                request_id=request_id,
-                auto_import=will_auto_import,
-            ),
-            db=db,
-        )
-        album_data.import_folder = dest
-        log_validation_result(album_data, bv_result, ctx.cfg, dest_path=dest)
-        logger.info(f"STAGED: {album_data.artist} - {album_data.title} "
-                    f"(scenario={bv_result.scenario}, "
-                    f"distance={bv_result.distance:.4f}) → {dest}")
-
-        dl_info = _build_download_info(album_data)
-        dl_info.validation_result = bv_result.to_json()
-        if album_data.download_spectral is not None:
-            dl_info.download_spectral = album_data.download_spectral
-            dl_info.current_spectral = album_data.current_spectral
-            dl_info.existing_min_bitrate = album_data.current_min_bitrate
-            dl_info.slskd_filetype = dl_info.filetype
-            dl_info.actual_filetype = dl_info.filetype
-        if will_auto_import:
-            assert request_id is not None, "pipeline request must have db_request_id"
-            assert pdb is not None, "auto-import path must hold a pipeline DB handle"
-            override_min_bitrate: int | None = None
-            try:
-                req = pdb.get_request(request_id)
-                if req:
-                    override_min_bitrate = compute_effective_override_bitrate(
-                        req.get("min_bitrate"),
-                        req.get("current_spectral_bitrate"),
-                        req.get("current_spectral_grade"),
-                    )
-            except Exception:
-                logger.debug("DB lookup failed for override-min-bitrate")
-
-            core_kwargs: dict[str, Any] = dict(
-                path=dest,
-                mb_release_id=album_data.mb_release_id or "",
-                request_id=request_id,
-                label=f"{album_data.artist} - {album_data.title}",
-                override_min_bitrate=override_min_bitrate,
-                target_format=album_data.db_target_format,
-                verified_lossless_target=ctx.cfg.verified_lossless_target,
-                beets_harness_path=ctx.cfg.beets_harness_path,
-                db=pdb,
-                dl_info=dl_info,
-                distance=bv_result.distance if bv_result.distance is not None else 0.0,
-                scenario=bv_result.scenario or "auto_import",
-                files=album_data.files,
-                cfg=ctx.cfg,
-                requeue_on_failure=True,
-                cooled_down_users=ctx.cooled_down_users,
-                source_dirs=_source_dirs_for_album(album_data),
-                candidate_import_job_id=import_job_id,
-                prevalidated_candidate_result=prevalidated_candidate_result,
-            )
-            if quality_gate_fn is not None:
-                core_kwargs["quality_gate_fn"] = quality_gate_fn
-            _dispatch = dispatch_fn if dispatch_fn is not None else dispatch_import_core
-            return _dispatch(**core_kwargs)
-        ctx.pipeline_db_source.mark_done(
-            album_data, bv_result, dest_path=dest, download_info=dl_info)
-        return None
-
-
-def _handle_rejected_result(album_data: GrabListEntry, bv_result: ValidationResult,
-                            staged_album: StagedAlbum,
-                            ctx: CratediggerContext,
-                            *,
-                            import_job_id: int | None = None) -> DispatchOutcome:
-    """Handle a rejected beets validation result."""
-    bv_result.source_dirs = _source_dirs_for_album(album_data)
-    failed_dest = move_failed_import_curated(
-        staged_album.current_path,
-        allowed_audio=tracked_audio_paths_for_downloads(album_data.files),
-        scenario=bv_result.scenario,
-    )
-    bv_result.failed_path = failed_dest
-    log_validation_result(album_data, bv_result, ctx.cfg)
-    usernames = set(f.username for f in album_data.files)
-    bv_result.denylisted_users = sorted(usernames)
-    dl_info = _build_download_info(album_data)
-    dl_info.validation_result = bv_result.to_json()
-    if album_data.download_spectral is not None:
-        dl_info.download_spectral = album_data.download_spectral
-        dl_info.current_spectral = album_data.current_spectral
-        dl_info.existing_min_bitrate = album_data.current_min_bitrate
-        dl_info.slskd_filetype = dl_info.filetype
-        dl_info.actual_filetype = dl_info.filetype
-
-    # Backfill search_filetype_override for pre-quality-gate albums stuck in loops
-    backfill_override = _compute_rejection_backfill(album_data, ctx)
-
-    download_log_id = ctx.pipeline_db_source.reject_and_requeue(
-        album_data,
-        bv_result,
-        usernames=usernames,
-        download_info=dl_info,
-        search_filetype_override=backfill_override,
-        cooled_down_users=ctx.cooled_down_users,
-    )
-    _run_post_rejection_wrong_match_cleanup(
-        ctx,
-        download_log_id,
-        scenario=bv_result.scenario,
-        import_job_id=import_job_id,
-    )
-    logger.warning(f"REJECTED: {album_data.artist} - {album_data.title} "
-                   f"(scenario={bv_result.scenario}, "
-                   f"distance={bv_result.distance}, "
-                   f"detail={bv_result.detail}) "
-                   f"| denylisted users: {', '.join(usernames)}")
-    scenario = bv_result.scenario or "validation_rejected"
-    detail = bv_result.detail or bv_result.error
-    message = f"Rejected: {scenario}"
-    if detail:
-        message = f"{message} - {detail}"
-    return DispatchOutcome(success=False, message=message)
-
-
-def _compute_rejection_backfill(album_data: GrabListEntry,
-                                ctx: CratediggerContext) -> str | None:
-    """Check if search_filetype_override should be backfilled on rejection.
-
-    Only fires when search_filetype_override is currently NULL and the on-disk state
-    is genuine + decent quality + not verified lossless.
-    """
-    request_id = album_data.db_request_id
-    if not request_id or not ctx.pipeline_db_source:
-        return None
-    if album_data.db_search_filetype_override:
-        return None
-    try:
-        db = ctx.pipeline_db_source._get_db()
-        req = db.get_request(request_id)
-        if not req or req.get("search_filetype_override"):
-            return None
-        from lib.beets_db import BeetsDB
-        with BeetsDB() as beets:
-            info = beets.get_album_info(
-                album_data.mb_release_id, ctx.cfg.quality_ranks)
-        if not info:
-            return None
-        override = rejection_backfill_override(
-            is_cbr=info.is_cbr,
-            min_bitrate_kbps=info.min_bitrate_kbps,
-            spectral_grade=req.get("current_spectral_grade"),
-            verified_lossless=bool(req.get("verified_lossless")),
-            cfg=ctx.cfg.quality_ranks,
-        )
-        if override:
-            logger.info(
-                f"BACKFILL: {album_data.artist} - {album_data.title} "
-                f"search_filetype_override=NULL → '{override}' "
-                f"(on-disk: {info.min_bitrate_kbps}kbps, cbr={info.is_cbr}, "
-                f"spectral={req.get('current_spectral_grade')})")
-        return override
-    except Exception:
-        logger.debug("BACKFILL: failed to check on-disk state", exc_info=True)
-        return None
 
 
 # === ActiveDownloadState building ===
@@ -1701,167 +197,6 @@ def build_active_download_state(
         ),
     )
 
-
-# === Transfer ID re-derivation ===
-
-def match_transfer_id(
-    downloads: dict[str, Any] | list[dict[str, Any]],
-    target_filename: str,
-    username: str | None = None,
-) -> str | None:
-    """Find the slskd transfer ID for a filename in slskd download responses.
-
-    downloads may be a single-user transfer dict or the list returned by
-    slskd.transfers.get_all_downloads(). When a list is provided, username
-    narrows the search to one peer.
-    Returns the transfer ID string, or None if not found.
-    """
-    transfer = match_transfer(downloads, target_filename, username=username)
-    if transfer is None:
-        return None
-    return transfer.get("id", "")
-
-
-def _parse_transfer_timestamp(value: Any) -> datetime:
-    """Parse slskd transfer timestamps for ordering duplicate snapshots."""
-    if not value:
-        return datetime.min.replace(tzinfo=timezone.utc)
-
-    text = str(value).replace("Z", "+00:00")
-    try:
-        parsed = datetime.fromisoformat(text)
-    except ValueError:
-        return datetime.min.replace(tzinfo=timezone.utc)
-
-    if parsed.tzinfo is None:
-        parsed = parsed.replace(tzinfo=timezone.utc)
-    return parsed.astimezone(timezone.utc)
-
-
-def _transfer_latest_timestamp(transfer: dict[str, Any]) -> datetime:
-    return max(
-        _parse_transfer_timestamp(transfer.get("endedAt")),
-        _parse_transfer_timestamp(transfer.get("startedAt")),
-        _parse_transfer_timestamp(transfer.get("enqueuedAt")),
-        _parse_transfer_timestamp(transfer.get("requestedAt")),
-    )
-
-
-def _transfer_priority(transfer: dict[str, Any]) -> tuple[int, int, datetime]:
-    """Rank duplicate transfer snapshots for the same username+filename.
-
-    Prefer active transfers over terminal ones. Among terminal snapshots,
-    prefer successful completions over cancelled/errored attempts, and then
-    pick the newest lifecycle timestamp.
-    """
-    state = str(transfer.get("state", ""))
-    is_terminal = state.startswith("Completed,")
-    is_success = state == "Completed, Succeeded"
-    latest_ts = _transfer_latest_timestamp(transfer)
-    return (0 if is_terminal else 1, 1 if is_success else 0, latest_ts)
-
-
-def _is_terminal_transfer_before(
-    transfer: dict[str, Any],
-    not_before: str | None,
-) -> bool:
-    if not_before is None:
-        return False
-    state = str(transfer.get("state", ""))
-    if not state.startswith("Completed,"):
-        return False
-    threshold = _parse_transfer_timestamp(not_before)
-    latest_ts = _transfer_latest_timestamp(transfer)
-    if latest_ts == datetime.min.replace(tzinfo=timezone.utc):
-        return False
-    return latest_ts < threshold
-
-
-def match_transfer(
-    downloads: dict[str, Any] | list[dict[str, Any]],
-    target_filename: str,
-    username: str | None = None,
-) -> dict[str, Any] | None:
-    """Find the best slskd transfer snapshot for a username+filename pair."""
-    groups = downloads if isinstance(downloads, list) else [downloads]
-    candidates: list[dict[str, Any]] = []
-    for group in groups:
-        if username is not None and group.get("username") not in (None, "", username):
-            continue
-        for directory in group.get("directories", []):
-            for slskd_file in directory.get("files", []):
-                if slskd_file.get("filename") == target_filename:
-                    candidates.append(slskd_file)
-
-    if not candidates:
-        return None
-    return max(candidates, key=_transfer_priority)
-
-
-def _get_all_downloads_snapshot(
-    slskd_client: Any,
-    *,
-    purpose: str,
-) -> list[dict[str, Any]] | None:
-    """Fetch the full slskd download snapshot via the bulk endpoint.
-
-    The username-scoped endpoint is unreliable for some valid peer names
-    containing spaces/punctuation, so monitoring code uses the bulk list and
-    matches locally instead.
-    """
-    try:
-        downloads = slskd_client.transfers.get_all_downloads(includeRemoved=True)
-    except Exception:
-        logger.warning(f"Failed to get all downloads for {purpose}", exc_info=True)
-        return None
-
-    if not isinstance(downloads, list):
-        logger.warning("Unexpected get_all_downloads() response type %s",
-                       type(downloads).__name__)
-        return None
-
-    return downloads
-
-
-def rederive_transfer_ids(
-    entry: GrabListEntry,
-    slskd_client: Any,
-    *,
-    snapshot: list[dict[str, Any]] | None = None,
-    not_before: str | None = None,
-) -> bool:
-    """Re-derive slskd transfer IDs for all files in a GrabListEntry.
-
-    Queries the slskd bulk download API once and matches by username+filename.
-    Updates file.id in-place. Files whose transfers have vanished keep id="".
-    When snapshot is provided, uses it instead of fetching from the API.
-    """
-    downloads = snapshot
-    if downloads is None:
-        downloads = _get_all_downloads_snapshot(
-            slskd_client,
-            purpose=f"transfer re-derivation for {entry.artist} - {entry.title}",
-        )
-    if downloads is None:
-        return False
-
-    for f in entry.files:
-        transfer = match_transfer(downloads, f.filename, username=f.username)
-        if transfer is not None and _is_terminal_transfer_before(
-            transfer,
-            not_before,
-        ):
-            transfer = None
-        if transfer is not None:
-            f.id = transfer.get("id", "")
-            state = str(transfer.get("state", ""))
-            if state.startswith("Completed,"):
-                f.status = dict(transfer)
-            else:
-                f.status = None
-        else:
-            logger.debug(f"Transfer not found for {f.filename} from {f.username}")
-    return True
 
 
 # === GrabListEntry reconstruction from DB ===
@@ -1928,6 +263,7 @@ def _restored_terminal_status(
         "state": last_state,
         "bytesTransferred": bytes_transferred,
     }
+
 
 
 # === Async download polling ===
