+++ scripts/importer.py
@@ -321,7 +321,8 @@ def execute_youtube_import_job(
     the staged audio manifest, so the rejection paths inside
     ``_handle_rejected_result`` find no peers to denylist.
     """
-    from lib.download import process_completed_album, reconstruct_grab_list_entry
+    from lib.download import reconstruct_grab_list_entry
+    from lib.download_processing import process_completed_album
 
     request_id = job.request_id
     if request_id is None:
